:recycle:  refactor: MessageType 수정 [#34]
